ai/requirements.txt
torch
torchvision
transformers>=4.38
datasets
tqdm
numpy
pillow
accelerate